# Conflicts:
#	Source/KORACustomEditors/Private/DataTool/TableRowConvertFunctionContainer.cpp
#	Source/KORACustomEditors/Public/DataTool/TableRowConvertFunctionContainer.h
#	Source/KORAProject/Data/CacheDataTable.cpp
#	Source/KORAProject/Data/GameDataType.h
#	Source/KORAProject/Game/KRGameInstance.cpp
#	Source/KORAProject/GameplayTag/KRAbilityTag.cpp
#	Source/KORAProject/GameplayTag/KRAbilityTag.h
#	Source/KORAProject/Inventory/Fragment/InventoryFragment_ModuleItem.cpp
#	Source/KORAProject/Inventory/Fragment/InventoryFragment_QuestItem.cpp
#	Source/KORAProject/Inventory/Fragment/InventoryFragment_SetStats.cpp
#	Source/KORAProject/SubSystem/KRDataTablesSubsystem.cpp
#	Source/KORAProject/SubSystem/KRDataTablesSubsystem.h

+++ Source/KORAProject/Inventory/Fragment/InventoryFragment_ModuleItem.cpp
@@ -39,7 +39,7 @@ void UInventoryFragment_ModuleItem::InitializeFromData(UKRInventoryItemInstance*
 		TargetWeaponSlotTag = KRTAG_ITEMTYPE_EQUIP_GUN;
 	}
 	
-	const FEquipAbilityDataStruct* AbilityData = DataSubsystem.GetData<FEquipAbilityDataStruct>(static_cast<uint32>(ModuleData->EquipAbilityID), TEXT("ModuleFragment"), false);
+	const FEquipAbilityDataStruct* AbilityData = DataSubsystem.GetData<FEquipAbilityDataStruct>((ModuleData->EquipAbilityID), TEXT("ModuleFragment"), false);
 
 	if (AbilityData)
 	{
